+++ collectors/HackerTarget.py
@@ -9,7 +9,7 @@ def init(domain):
 	print(colored("[*]-Searching HackerTarget...", "yellow"))
 
 	url = "https://api.hackertarget.com/hostsearch/?q={0}".format(quote(domain))
-	headers = {"user-agent": "Mozilla/5.0 (Macintosh; Intel Mac OS X 10.10; rv:52.0) Gecko/20100101 Firefox/52.0"}
+	headers = {"user-agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/88.0.4324.150 Safari/537.36"}
 
 	try:
 		response = requests.get(url, headers=headers).text