+++ collectors/Spyse.py
@@ -17,7 +17,7 @@ def init(domain):
 		print("  \__", colored("No Spyse API token configured", "red"))
 		return []
 
-	headers = {"user-agent": "Mozilla/5.0 (Macintosh; Intel Mac OS X 10.10; rv:52.0) Gecko/20100101 Firefox/52.0"}
+	headers = {"user-agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/88.0.4324.150 Safari/537.36"}
 	url = "https://api.spyse.com/v1/subdomains?api_token={0}&domain={1}&page=".format(SPYSE_API_TOKEN, domain)
 
 	try: